refactor(server): split websocket production helpers
Move the remaining large websocket production helpers behind focused child modules while preserving the existing XMPP behavior. PubSub fan-out keeps the public publish entry point but moves the XEP-0163 CAPS fan-out pass; stream management keeps nonza dispatch in place and moves post-registration resume finalization; disco#info now routes to target-specific modules without changing branch order or feature advertisements.

Constraint: One PR for the remaining issue #350 production cleanup

Constraint: Refactor-only; no XMPP/XEP wire-shape, namespace, or advertised-feature changes

Rejected: Rewrite the handlers into a new router abstraction | would widen the behavioral surface of a cleanup PR

Rejected: Split every remaining websocket helper file | outside the agreed three-file cleanup scope

Confidence: high

Scope-risk: moderate

Directive: Keep future XEP behavior changes in dedicated PRs with dedicated XEP tests, not in this cleanup split

Tested: cargo test -p waddle-server pubsub --lib

Tested: cargo test -p waddle-server stream_management --lib

Tested: cargo test -p waddle-server disco_info --lib

Tested: cargo test -p waddle-server handle_iq_cross_user_pep_disco_resolves_session_backed_accounts --lib

Tested: cargo test -p waddle-server server::routes::websocket --lib

Tested: cargo test -p waddle-server

Tested: cargo clippy -p waddle-server --tests -- -D warnings

Tested: git diff --cached --check
